Restructuring and cleaning/commenting files in Components/Views (#463)
* File Restructure in Components/Views

* Announcements View Pages cleanup and comments

* Hame Page cleanup

* MyTeamPage stuff cleanup

* Reschedule Cleanup

* Schedule Page Cleanup

* View cleanup and comments

* fix comments

* Add constants file instead of repeating code for colours constants

* Structural and comments updates

* Added a new constant and removed repeated code
