+++ src/frontend/src/api/announcements.js
@@ -1,6 +1,5 @@
 // Announcement API functions: Fetch, create, edit, and delete announcements.
-
-const REACT_APP_API_BASE_URL = "http://localhost:3001/api"; 
+import { REACT_APP_API_BASE_URL } from "../utils/Constants";
 
 // Fetch all announcements
 export async function getAnnouncements() {
