+++ src/frontend/src/components/Tables/CommissionerContactInfo.jsx
@@ -9,14 +9,7 @@ import {
   Paper
 } from "@mui/material";
 import ContactInfoTable from "../Tables/ContactInfoTable";
-
-// McMaster colours - AI Generated
-const MCMASTER_COLOURS = {
-  maroon: '#7A003C',
-  grey: '#5E6A71',
-  gold: '#FDBF57',
-  lightGrey: '#F5F5F5',
-};
+import { MCMASTER_COLOURS } from "../../utils/Constants";
 
 // CommissionerContactInfo: Displays a contact information table for the selected season.
 export default function CommissionerContactInfo({ seasons }) {
