+++ src/frontend/src/components/Tables/RosterTable.jsx
@@ -9,14 +9,7 @@ import {
   TableRow,
   Typography,
 } from "@mui/material";
-
-// McMaster colours - AI generated
-const MCMASTER_COLOURS = {
-  maroon: '#7A003C',
-  grey: '#5E6A71',
-  gold: '#FDBF57',
-  lightGrey: '#F5F5F5',
-};
+import { MCMASTER_COLOURS } from "../../utils/Constants";
 
 // RosterTable: Displays a table of players in a team with their contact information.
 export default function RosterTable(props) {
