+++ src/frontend/src/context/AuthContext.js
@@ -1,5 +1,6 @@
 import { createContext } from "react";
 
+// AuthContext: Context for managing authentication state
 export const AuthContext = createContext({
   isLoggedIn: false,
   playerId: null,
