+++ src/frontend/src/deprecated_dummy_data/schedule.json
