+++ src/frontend/src/deprecated_dummy_data/seasons.json
