+++ src/frontend/src/deprecated_dummy_data/standings.json
