+++ src/frontend/src/utils/Constants.js
@@ -1,0 +1,8 @@
+export const MCMASTER_COLOURS = {
+  maroon: "#7A003C",
+  grey: "#5E6A71",
+  gold: "#FDBF57",
+  lightGrey: "#F5F5F5",
+};
+
+export const REACT_APP_API_BASE_URL = "http://localhost:3001/api";
