+++ src/frontend/src/views/Announcements/AnnouncementsPage.jsx
@@ -1,21 +1,15 @@
 import React, { useState, useEffect } from "react";
 import { Box, Container, Typography, Button } from "@mui/material";
 import { useLocation, useNavigate } from "react-router-dom";
-import NavBar from "../components/NavBar";
-import { getAnnouncements } from "../api/announcements";
-import PastAnnouncementsSection from "../components/Cards/PastAnnouncements";
+import NavBar from "../../components/NavBar";
+import { getAnnouncements } from "../../api/announcements";
+import PastAnnouncementsSection from "../../components/Cards/PastAnnouncements";
 import EditIcon from "@mui/icons-material/Edit";
-import { useAuth } from "../hooks/AuthProvider";
-import { getPlayerById } from "../api/player";
+import { useAuth } from "../../hooks/AuthProvider";
+import { getPlayerById } from "../../api/player";
+import { MCMASTER_COLOURS } from "../../utils/Constants.js";
 
-// McMaster colours - AI Generated
-const MCMASTER_COLOURS = {
-  maroon: '#7A003C',
-  grey: '#5E6A71',
-  gold: '#FDBF57',
-  lightGrey: '#F5F5F5',
-};
-
+// AnnouncementPage: Displays the main announcement and past announcements.
 export default function AnnouncementPage() {
   const navigate = useNavigate();
   const location = useLocation();
@@ -25,6 +19,7 @@ export default function AnnouncementPage() {
   const auth = useAuth();
   const [player, setPlayer] = useState(null);
 
+  // fetches the current player data
   useEffect(() => {
     const fetchPlayer = async () => {
       if (!auth.playerId) return;
@@ -39,6 +34,7 @@ export default function AnnouncementPage() {
     fetchPlayer();
   }, [auth.playerId]);
 
+  // fetches the announcements data
   useEffect(() => {
     const fetchAnnouncements = async () => {
       try {
