+++ src/frontend/src/views/Announcements/CreateAnnouncementsPage.jsx
@@ -1,21 +1,21 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { Container, Typography, Box } from "@mui/material";
-import NavBar from "../components/NavBar";
-import AnnouncementForm from "../components/Forms/AnnouncementForm";
-import { createAnnouncement } from "../api/announcements";
+import NavBar from "../../components/NavBar";
+import AnnouncementForm from "../../components/Forms/AnnouncementForm";
+import { createAnnouncement } from "../../api/announcements";
+import { MCMASTER_COLOURS } from "../../utils/Constants.js";
 
-// McMaster colours - AI Generated
-const MCMASTER_COLOURS = {
-  maroon: '#7A003C',
-  grey: '#5E6A71',
-  gold: '#FDBF57',
-  lightGrey: '#F5F5F5',
-};
 
+// CreateAnnouncementPage: Redirect Page for creating a new announcement (commissioner).
 export default function CreateAnnouncementPage() {
   const navigate = useNavigate();
 
+  ```  
+  Function to handle the creation of a new announcement
+  It takes the title and content as parameters and calls the createAnnouncement API function,
+  then navigates back to the announcements page.
+  ```
   const handleCreate = async ({ title, content }) => {
     try {
       await createAnnouncement(title, content);
