+++ src/frontend/src/views/Announcements/EditAnnouncementsPage.jsx
@@ -1,19 +1,14 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import { Container, Typography, Box, CircularProgress } from "@mui/material";
-import NavBar from "../components/NavBar";
-import AnnouncementForm from "../components/Forms/AnnouncementForm";
-import { getAnnouncementById, editAnnouncement, deleteAnnouncement } from "../api/announcements";
+import NavBar from "../../components/NavBar";
+import AnnouncementForm from "../../components/Forms/AnnouncementForm";
+import { getAnnouncementById, editAnnouncement, deleteAnnouncement } from "../../api/announcements";
+import { MCMASTER_COLOURS } from "../../utils/Constants.js";
 
-// McMaster colours - AI Generated
-const MCMASTER_COLOURS = {
-  maroon: '#7A003C',
-  grey: '#5E6A71',
-  gold: '#FDBF57',
-  lightGrey: '#F5F5F5',
-};
 
 // AI Generated - Ombre bar styling and gradient effects
+// EditAnnouncementPage: Redirect Page for editing an existing announcement (commissioner).
 export default function EditAnnouncementPage() {
   const { id } = useParams();
   const navigate = useNavigate();
@@ -21,6 +16,7 @@ export default function EditAnnouncementPage() {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
 
+  // Fetch the announcement by ID when the component mounts
   useEffect(() => {
     const fetchAnnouncement = async () => {
       try {
@@ -36,6 +32,7 @@ export default function EditAnnouncementPage() {
     fetchAnnouncement();
   }, [id]);
 
+  // Function to handle the edit onclick - sends edit request to the API
   const handleEdit = async ({ title, content }) => {
     try {
       await editAnnouncement(id, title, content);
@@ -46,6 +43,7 @@ export default function EditAnnouncementPage() {
     }
   };
 
+  // Function to handle the delete button click - sends delete request to the API
   const handleDelete = async () => {
     if (window.confirm("Are you sure you want to delete this announcement?")) {
       try {
