+++ src/frontend/src/views/Deprecated/SchedulePage.jsx
@@ -12,11 +12,12 @@ import TabContext from "@mui/lab/TabContext";
 import TabList from "@mui/lab/TabList";
 import TabPanel from "@mui/lab/TabPanel";
 
-import NavBar from "../components/NavBar";
-import schedule from "../data/schedule.json";
-import seasons from "../data/seasons.json";
-import ScheduleTable from "../components/Tables/ScheduleTable";
+import NavBar from "../../components/NavBar";
+import schedule from "../../data/schedule.json";
+import seasons from "../../data/seasons.json";
+import ScheduleTable from "../../components/Tables/ScheduleTable";
 
+// Deprecated Schedule Page , replaced by ScheduleView
 export default function SchedulePage() {
   //   const [selectedSeason, setSelectedSeason] = useState("2025");
   //   const [selectedDivision, setSelectedDivision] = useState("Division A");
