+++ src/frontend/src/views/MyTeam/PlayersPage.jsx
@@ -1,10 +1,11 @@
 import { Typography, Container } from "@mui/material";
-import NavBar from "../components/NavBar";
-import PlayerTable from "../components/Tables/PlayerTable";
-import { allPlayers } from "../api/player";
+import NavBar from "../../components/NavBar";
+import PlayerTable from "../../components/Tables/PlayerTable";
+import { allPlayers } from "../../api/player";
 import { useState, useEffect } from "react";
-import LoadingOverlay from "../components/LoadingOverlay";
+import LoadingOverlay from "../../components/LoadingOverlay";
 
+// Page for searching and inviting players to your team as a captain
 export default function Players() {
   const [loading, setLoading] = useState(true);
   const [players, setPlayers] = useState([]);
