+++ src/frontend/src/views/ProfilePage.jsx
@@ -2,7 +2,6 @@ import React, { useState, useEffect } from "react";
 import NavBar from "../components/NavBar";
 import { getPlayerById } from "../api/player";
 import { useAuth } from "../hooks/AuthProvider";
-import { acceptInvite } from "../api/player";
 import { updatePlayerInfo } from "../api/player";
 
 import {
@@ -26,6 +25,7 @@ const ProfileContainer = styled(Box)(({ theme }) => ({
   marginBottom: theme.spacing(4),
 }));
 
+// ProfilePage component: Displays and allows editing of user profile information
 export default function ProfilePage() {
   const auth = useAuth();
