+++ src/frontend/src/views/Reschedule/ReschedulePopup.jsx
@@ -1,48 +1,9 @@
 import React, { useState, useEffect } from "react";
-import { createRescheduleRequest } from "../api/reschedule-requests";
+import { createRescheduleRequest } from "../../api/reschedule-requests";
+import { getPopupWeekDates, getPopupWeekRange } from "../../utils/Formatting";
 
-// Utility functions
-function getLocalISODate(date) {
-  date.setHours(12, 0, 0, 0);
-  const offset = date.getTimezoneOffset();
-  const localDate = new Date(date.getTime() - offset * 60 * 1000);
-  return localDate.toISOString().split("T")[0];
-}
-
-const getPopupWeekDates = (date) => {
-  let monday = new Date(date);
-  const day = monday.getDay();
-  if (day === 0) {
-    // If Sunday, jump to Monday
-    monday.setDate(monday.getDate() + 1);
-  } else {
-    // Otherwise, back up to Monday
-    monday.setDate(monday.getDate() - (day - 1));
-  }
-  // Return 5 days (Monday to Friday)
-  return [...Array(5)].map((_, i) => {
-    let dayDate = new Date(monday);
-    dayDate.setDate(monday.getDate() + i);
-    const fullDate = getLocalISODate(dayDate);
-    return {
-      day: dayDate.toLocaleDateString("en-US", { weekday: "long" }),
-      date: dayDate.toLocaleDateString("en-US", { day: "numeric", month: "short" }),
-      fullDate,
-    };
-  });
-};
-
-function getPopupWeekRange(date) {
-  let startOfWeek = new Date(date);
-  startOfWeek.setDate(startOfWeek.getDate() - startOfWeek.getDay());
-  let endOfWeek = new Date(startOfWeek);
-  endOfWeek.setDate(startOfWeek.getDate() + 6);
-
-  const startStr = startOfWeek.toLocaleDateString("en-US", { month: "long", day: "numeric" });
-  const endStr = endOfWeek.toLocaleDateString("en-US", { month: "long", day: "numeric", year: "numeric" });
-  return `${startStr} - ${endStr}`;
-}
-
+// Popup modal for rescheduling a match, showing available timeslots and allows
+// the user to select new timeslots for the match and scroll through the week
 export const ReschedulePopup = ({ selectedDate, selectedMatch, availableTimeslots, player, onClose }) => {
   const [popupDate, setPopupDate] = useState(new Date(selectedDate));
   const [loading, setLoading] = useState(false);
@@ -55,9 +16,7 @@ export const ReschedulePopup = ({ selectedDate, selectedMatch, availableTimeslot
   const popupWeekRange = getPopupWeekRange(popupDate);
   const [newSlots, setNewSlots] = useState([]);
 
-
   // Fetch available slots from backend
-
   const handleWeekNav = (direction) => {
     let newDate = new Date(popupDate);
     newDate.setDate(newDate.getDate() + (direction === "next" ? 7 : -7));
